// src/app/state.types.ts
// Generated by https://quicktype.io

export interface TopLevel {
  users: User[];
}

export interface User {
  id: string;
  username: string;
  lastSelectedChronicle: string;
  chronicles: Chronicle[];
}

export interface Chronicle {
  id: string;
  name: string;
  lastSelectedCharacter: null | string;
  clans: Clan[];
}

export interface Clan {
  name: string;
  characters: Character[];
}

export interface Character {
  id: string;
  name: string;
  imageUrl: string;
  nature: string;
  demeanor: string;
  generation: number;
  concept: string;
  lastSelectedNote: null | string;
  notes: Note[];
}

export interface Note {
  id: string;
  title: string;
  subtitle: string;
  content: string;
}